+++ src/Cli/Framework.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Watson\Cli;
+
+enum Framework: string
+{
+    case Symfony = 'symfony';
+    case Laravel = 'laravel';
+}